refactor(recordings): own the terminal run-finished event instead of importing Factory Runtime
Recordings owns the canonical Factory Event ledger, yet its lifecycle runtime
recorder imported pkg/services/factory_runtime to name the events it writes
into that ledger. The ledger owner depended on a consumer of the ledger for
its own vocabulary. That single import is the whole recordings ->
factory_runtime back-edge, and because Factory Runtime's root already imports
Recordings in three files, it also closed a genuine mutual loop between the
two services rather than a one-way edge.

The import turned out to carry two very different things.

Four of the six uses were the FactoryEvent type, which is not Factory
Runtime's type at all: factory_runtime.FactoryEvent, recordings.FactoryEvent
and factorydefinitions.FactoryEvent are three `=` aliases of one struct. Those
sites now spell the type through Recordings' own root, which is a rename with
no type identity, interface satisfaction, or serialization change.

The sixth use was real. RunFinishedFactoryEvent builds the terminal
RUN_RESPONSE entry, and it moves to pkg/services/recordings, where the rest of
the ledger vocabulary already lives -- Recordings root already re-exported
every alias the builder needs, and Recordings already recognized this event
identity in its replay and event-history paths. Factory Runtime's copy is
deleted rather than left as a forwarding alias, because the only non-test
caller was the Recordings recorder; keeping a wrapper would preserve the
coupling in the docs while hiding it from the checker.

The builder's test moves with it. The recorder's own boundary test keeps its
end-to-end claim -- a runtime-produced event and the appended terminal event
both survive into the portable artifact with their identity, kind and payload
intact -- restated in Recordings vocabulary. Its two leftover import-path
constants are removed: they belonged to an import-shape guard that had already
been deleted, and the claim is now enforced repo-wide by the cycle ratchet.

Measured cross-service minimum feedback arc weight falls from 14 to 13 and
recordings -> factory_runtime leaves the cut set.
