+++ pkg/services/factory_runtime/recording_event_contracts.go
@@ -1,15 +1,14 @@
 package factory
 
-import (
-	"encoding/json"
-	"time"
+import interfaces "github.com/portpowered/infinite-you/pkg/services/factory_definitions"
 
-	interfaces "github.com/portpowered/infinite-you/pkg/services/factory_definitions"
-)
-
-// Runtime-owned Factory Event vocabulary published at the Factory Runtime root.
-// Recordings lifecycle recorder edges consume these aliases rather than nested
-// Runtime implementation packages or loose helper paths.
+// Runtime-owned Factory Event vocabulary published at the Factory Runtime root
+// so Runtime callers name event shapes here rather than through nested Runtime
+// implementation packages or loose helper paths.
+//
+// These are aliases of the authored definition vocabulary, not a second event
+// taxonomy. The terminal run-finished event is owned by pkg/services/recordings
+// with the rest of the canonical Factory Event ledger.
 type (
 	FactoryEvent            = interfaces.FactoryEvent
 	FactoryEventContext     = interfaces.FactoryEventContext
@@ -25,35 +24,3 @@ const (
 	FactoryEventTypeWorkRequest = interfaces.FactoryEventTypeWorkRequest
 	FactoryStateCompleted       = interfaces.FactoryStateCompleted
 )
-
-// RunFinishedFactoryEventID is the stable identity for terminal run completion
-// events emitted through the Runtime root recording vocabulary.
-const RunFinishedFactoryEventID = "factory-event/run-finished"
-
-// RunFinishedFactoryEvent constructs the terminal RUN_RESPONSE event emitted
-// when a Factory Runtime instance completes. Recordings lifecycle recorder
-// edges publish this shape through the Runtime root contract.
-func RunFinishedFactoryEvent(startedAt, finishedAt time.Time) FactoryEvent {
-	state := FactoryStateCompleted
-	startedAtUTC := startedAt.UTC()
-	finishedAtUTC := finishedAt.UTC()
-	payload, err := json.Marshal(RunResponseEventPayload{
-		State: &state,
-		WallClock: &RunEventWallClock{
-			StartedAt:  &startedAtUTC,
-			FinishedAt: &finishedAtUTC,
-		},
-	})
-	if err != nil {
-		payload = json.RawMessage(`{}`)
-	}
-	return FactoryEvent{
-		Id:            RunFinishedFactoryEventID,
-		SchemaVersion: FactoryEventSchemaVersionV1,
-		Type:          FactoryEventTypeRunResponse,
-		Context: FactoryEventContext{
-			EventTime: finishedAtUTC,
-		},
-		Payload: payload,
-	}
-}
